experimental[patch]: refactor rl chain structure
