+++ libs/community/langchain_community/embeddings/infinity.py
@@ -7,6 +7,8 @@ from typing import Any, Callable, Dict, List, Optional, Tuple
 import aiohttp
 import numpy as np
 import requests
+
+import langchain_experimental.rl_chain.helpers
 from langchain_core.embeddings import Embeddings
 from langchain_core.pydantic_v1 import BaseModel, root_validator
 from langchain_core.utils import get_from_dict_or_env
@@ -69,7 +71,7 @@ class InfinityEmbeddings(BaseModel, Embeddings):
         Returns:
             List of embeddings, one for each text.
         """
-        embeddings = self.client.embed(
+        embeddings = langchain_experimental.rl_chain.helpers.embed(
             model=self.model,
             texts=texts,
         )
