+++ libs/community/langchain_community/embeddings/llamacpp.py
@@ -1,5 +1,6 @@
 from typing import Any, Dict, List, Optional
 
+import langchain_experimental.rl_chain.helpers
 from langchain_core.embeddings import Embeddings
 from langchain_core.pydantic_v1 import BaseModel, Field, root_validator
 
@@ -108,7 +109,7 @@ class LlamaCppEmbeddings(BaseModel, Embeddings):
         Returns:
             List of embeddings, one for each text.
         """
-        embeddings = [self.client.embed(text) for text in texts]
+        embeddings = [langchain_experimental.rl_chain.helpers.embed(text) for text in texts]
         return [list(map(float, e)) for e in embeddings]
 
     def embed_query(self, text: str) -> List[float]:
@@ -120,5 +121,5 @@ class LlamaCppEmbeddings(BaseModel, Embeddings):
         Returns:
             Embeddings for the text.
         """
-        embedding = self.client.embed(text)
+        embedding = langchain_experimental.rl_chain.helpers.embed(text)
         return list(map(float, embedding))
