+++ libs/community/langchain_community/embeddings/textembed.py
@@ -16,6 +16,8 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 import aiohttp
 import numpy as np
 import requests
+
+import langchain_experimental.rl_chain.helpers
 from langchain_core.embeddings import Embeddings
 from langchain_core.pydantic_v1 import BaseModel, root_validator
 from langchain_core.utils import get_from_dict_or_env
@@ -89,7 +91,7 @@ class TextEmbedEmbeddings(BaseModel, Embeddings):
         Returns:
             List[List[float]]: List of embeddings, one for each text.
         """
-        embeddings = self.client.embed(
+        embeddings = langchain_experimental.rl_chain.helpers.embed(
             model=self.model,
             texts=texts,
         )
