+++ libs/experimental/langchain_experimental/rl_chain/__init__.py
@@ -19,9 +19,8 @@ from langchain_experimental.rl_chain.base import (
     SelectionScorer,
     ToSelectFrom,
     VwPolicy,
-    embed,
-    stringify_embedding,
 )
+from langchain_experimental.rl_chain.helpers import embed, stringify_embedding
 from langchain_experimental.rl_chain.pick_best_chain import (
     PickBest,
     PickBestEvent,
