+++ libs/experimental/tests/unit_tests/rl_chain/test_pick_best_text_embedder.py
@@ -2,6 +2,7 @@ import pytest
 from test_utils import MockEncoder
 
 import langchain_experimental.rl_chain.base as rl_chain
+import langchain_experimental.rl_chain.helpers
 import langchain_experimental.rl_chain.pick_best_chain as pick_best_chain
 
 encoded_keyword = "[encoded]"
@@ -92,12 +93,20 @@ def test_pickbest_textembedder_w_full_label_w_emb() -> None:
     str1 = "0"
     str2 = "1"
     str3 = "2"
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
-    encoded_str2 = rl_chain.stringify_embedding(list(encoded_keyword + str2))
-    encoded_str3 = rl_chain.stringify_embedding(list(encoded_keyword + str3))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
+    encoded_str2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str2)
+    )
+    encoded_str3 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str3)
+    )
 
     ctx_str_1 = "context1"
-    encoded_ctx_str_1 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_1))
+    encoded_ctx_str_1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_1)
+    )
 
     named_actions = {"action1": rl_chain.Embed([str1, str2, str3])}
     context = {"context": rl_chain.Embed(ctx_str_1)}
@@ -118,12 +127,20 @@ def test_pickbest_textembedder_w_full_label_w_embed_and_keep() -> None:
     str1 = "0"
     str2 = "1"
     str3 = "2"
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
-    encoded_str2 = rl_chain.stringify_embedding(list(encoded_keyword + str2))
-    encoded_str3 = rl_chain.stringify_embedding(list(encoded_keyword + str3))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
+    encoded_str2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str2)
+    )
+    encoded_str3 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str3)
+    )
 
     ctx_str_1 = "context1"
-    encoded_ctx_str_1 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_1))
+    encoded_ctx_str_1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_1)
+    )
 
     named_actions = {"action1": rl_chain.EmbedAndKeep([str1, str2, str3])}
     context = {"context": rl_chain.EmbedAndKeep(ctx_str_1)}
@@ -192,14 +209,24 @@ def test_pickbest_textembedder_more_namespaces_w_full_label_w_full_emb() -> None
     str1 = "0"
     str2 = "1"
     str3 = "2"
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
-    encoded_str2 = rl_chain.stringify_embedding(list(encoded_keyword + str2))
-    encoded_str3 = rl_chain.stringify_embedding(list(encoded_keyword + str3))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
+    encoded_str2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str2)
+    )
+    encoded_str3 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str3)
+    )
 
     ctx_str_1 = "context1"
     ctx_str_2 = "context2"
-    encoded_ctx_str_1 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_1))
-    encoded_ctx_str_2 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_2))
+    encoded_ctx_str_1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_1)
+    )
+    encoded_ctx_str_2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_2)
+    )
 
     named_actions = {"action1": rl_chain.Embed([{"a": str1, "b": str1}, str2, str3])}
     context = {
@@ -227,14 +254,24 @@ def test_pickbest_textembedder_more_namespaces_w_full_label_w_full_embed_and_kee
     str1 = "0"
     str2 = "1"
     str3 = "2"
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
-    encoded_str2 = rl_chain.stringify_embedding(list(encoded_keyword + str2))
-    encoded_str3 = rl_chain.stringify_embedding(list(encoded_keyword + str3))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
+    encoded_str2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str2)
+    )
+    encoded_str3 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str3)
+    )
 
     ctx_str_1 = "context1"
     ctx_str_2 = "context2"
-    encoded_ctx_str_1 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_1))
-    encoded_ctx_str_2 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_2))
+    encoded_ctx_str_1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_1)
+    )
+    encoded_ctx_str_2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_2)
+    )
 
     named_actions = {
         "action1": rl_chain.EmbedAndKeep([{"a": str1, "b": str1}, str2, str3])
@@ -262,12 +299,18 @@ def test_pickbest_textembedder_more_namespaces_w_full_label_w_partial_emb() -> N
     str1 = "0"
     str2 = "1"
     str3 = "2"
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
-    encoded_str3 = rl_chain.stringify_embedding(list(encoded_keyword + str3))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
+    encoded_str3 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str3)
+    )
 
     ctx_str_1 = "context1"
     ctx_str_2 = "context2"
-    encoded_ctx_str_2 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_2))
+    encoded_ctx_str_2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_2)
+    )
 
     named_actions = {
         "action1": [
@@ -296,12 +339,18 @@ def test_pickbest_textembedder_more_namespaces_w_full_label_w_partial_emakeep() 
     str1 = "0"
     str2 = "1"
     str3 = "2"
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
-    encoded_str3 = rl_chain.stringify_embedding(list(encoded_keyword + str3))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
+    encoded_str3 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str3)
+    )
 
     ctx_str_1 = "context1"
     ctx_str_2 = "context2"
-    encoded_ctx_str_2 = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str_2))
+    encoded_ctx_str_2 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str_2)
+    )
 
     named_actions = {
         "action1": [
@@ -331,11 +380,15 @@ def test_raw_features_underscored() -> None:
     )
     str1 = "this is a long string"
     str1_underscored = str1.replace(" ", "_")
-    encoded_str1 = rl_chain.stringify_embedding(list(encoded_keyword + str1))
+    encoded_str1 = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + str1)
+    )
 
     ctx_str = "this is a long context"
     ctx_str_underscored = ctx_str.replace(" ", "_")
-    encoded_ctx_str = rl_chain.stringify_embedding(list(encoded_keyword + ctx_str))
+    encoded_ctx_str = langchain_experimental.rl_chain.helpers.stringify_embedding(
+        list(encoded_keyword + ctx_str)
+    )
 
     # No embeddings
     named_actions = {"action": [str1]}
