+++ libs/partners/ai21/langchain_ai21/embeddings.py
@@ -2,6 +2,8 @@ from itertools import islice
 from typing import Any, Iterator, List, Optional
 
 from ai21.models import EmbedType
+
+import langchain_experimental.rl_chain.helpers
 from langchain_core.embeddings import Embeddings
 
 from langchain_ai21.ai21_base import AI21Base
@@ -113,7 +115,7 @@ class AI21Embeddings(Embeddings, AI21Base):
     ) -> List[List[float]]:
         chunks = _split_texts_into_batches(texts, batch_size)
         responses = [
-            self.client.embed.create(
+            langchain_experimental.rl_chain.helpers.embed.create(
                 texts=chunk,
                 type=embed_type,
                 **kwargs,
