+++ libs/partners/voyageai/langchain_voyageai/embeddings.py
@@ -3,6 +3,8 @@ import os
 from typing import Iterable, List, Optional
 
 import voyageai  # type: ignore
+
+import langchain_experimental.rl_chain.helpers
 from langchain_core.embeddings import Embeddings
 from langchain_core.pydantic_v1 import (
     BaseModel,
@@ -87,7 +89,7 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
         _iter = self._get_batch_iterator(texts)
         for i in _iter:
             embeddings.extend(
-                self._client.embed(
+                langchain_experimental.rl_chain.helpers.embed(
                     texts[i : i + self.batch_size],
                     model=self.model,
                     input_type="document",
@@ -99,7 +101,7 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
 
     def embed_query(self, text: str) -> List[float]:
         """Embed query text."""
-        return self._client.embed(
+        return langchain_experimental.rl_chain.helpers.embed(
             [text], model=self.model, input_type="query", truncation=self.truncation
         ).embeddings[0]
 
@@ -108,7 +110,7 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
 
         _iter = self._get_batch_iterator(texts)
         for i in _iter:
-            r = await self._aclient.embed(
+            r = await langchain_experimental.rl_chain.helpers.embed(
                 texts[i : i + self.batch_size],
                 model=self.model,
                 input_type="document",
@@ -119,7 +121,7 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
         return embeddings
 
     async def aembed_query(self, text: str) -> List[float]:
-        r = await self._aclient.embed(
+        r = await langchain_experimental.rl_chain.helpers.embed(
             [text],
             model=self.model,
             input_type="query",
